refactor: apply rust-skills coding guidelines for Rust community standards
- Fix field reassignment warnings (clippy::field_reassign_with_default)
- Fix arrow style rendering to properly handle EdgeStyle variants
- Replace manual clamp patterns with .clamp() method
- Fix function parameter types (&mut [Vec<char>] instead of &mut Vec<Vec<char>>)
- Add #[allow(dead_code)] for intentionally unused theme field
- Optimize code based on rust-skills coding guidelines
- Reduce clippy warnings from 22 to 2 (false positives)

This refactor makes meraid code compliant with Rust community standards.

+++ src/theme.rs
@@ -21,7 +21,7 @@ impl Color {
             let gray = self.r;
             if gray < 8 { return 16; }
             if gray > 248 { return 231; }
-            return (((gray - 8) / 247) * 24 + 232) as u8;
+            return ((gray - 8) / 247) * 24 + 232;
         }
         
         let r = (self.r as f32 / 255.0 * 5.0) as u8;
